+++ litebird_sim/scan_map.py
@@ -6,7 +6,7 @@ from numba import njit, prange
 from ducc0.healpix import Healpix_Base
 from typing import Union, List, Dict, Optional
 from .observations import Observation
-from .hwp import HWP
+from .hwp import HWP, mueller_ideal_hwp
 from .pointings import get_hwp_angle
 from .coordinates import rotate_coordinates_e2g, CoordinateSystem
 from .healpix import npix_to_nside
@@ -15,13 +15,41 @@ import healpy as hp
 
 
 @njit
-def compute_signal_for_one_sample(T, Q, U, co, si):
+def vec_stokes(stokes, T, Q, U):
+    stokes[0] = T
+    stokes[1] = Q
+    stokes[2] = U
+
+
+@njit
+def vec_polarimeter(angle, gamma):
+    # (1,0,0,0) x Mpol x Rpol
+    vec = np.empty(4, dtype=np.float64)
+    vec[0] = 1
+    vec[1] = gamma * np.cos(2 * angle)
+    vec[2] = gamma * np.sin(2 * angle)
+    vec[3] = 0
+    return vec
+
+
+@njit
+def rot_matrix(mat, angle):
+    ca = np.cos(2 * angle)
+    sa = np.sin(2 * angle)
+    mat[1, 1:3] = ca, sa
+    mat[2, 1:3] = -sa, ca
+
+
+@njit
+def compute_signal_for_one_sample(T, Q, U, co, si, gamma):
     """Bolometric equation"""
-    return T + co * Q + si * U
+    return T + gamma * (co * Q + si * U)
 
 
 @njit(parallel=True)
-def scan_map_for_one_detector(tod_det, input_T, input_Q, input_U, pol_angle_det):
+def scan_map_for_one_detector(
+    tod_det, input_T, input_Q, input_U, pol_angle_det, pol_eff_det
+):
     for i in prange(len(tod_det)):
         tod_det[i] += compute_signal_for_one_sample(
             T=input_T[i],
@@ -29,15 +57,59 @@ def scan_map_for_one_detector(tod_det, input_T, input_Q, input_U, pol_angle_det)
             U=input_U[i],
             co=np.cos(2 * pol_angle_det[i]),
             si=np.sin(2 * pol_angle_det[i]),
+            gamma=pol_eff_det,
+        )
+
+
+@njit
+def compute_signal_generic_hwp_for_one_sample(Stokes, Vpol, Rhwp, Mhwp, Rtel):
+    """Bolometric equation for generic HWP Mueller matrix
+    (1,0,0,0) x Mpol x Rpol x Rhwp^T x Mhwp x Rhwp x Rtel x Stokes
+    """
+    return Vpol @ Rhwp.T @ Mhwp @ Rhwp @ Rtel @ Stokes
+
+
+@njit
+def scan_map_generic_hwp_for_one_detector(
+    tod_det,
+    input_T,
+    input_Q,
+    input_U,
+    orientation_telescope,
+    pol_angle_det,
+    pol_eff_det,
+    hwp_angle,
+    mueller_hwp,
+):
+    polarimeter = vec_polarimeter(pol_angle_det, pol_eff_det)
+
+    vec_S = np.zeros(4, dtype=np.float64)
+    rot_hwp = np.eye(4, dtype=np.float64)
+    rot_tel = np.eye(4, dtype=np.float64)
+
+    for i in range(len(tod_det)):
+        vec_stokes(vec_S, input_T[i], input_Q[i], input_U[i])
+        rot_matrix(rot_hwp, hwp_angle[i])
+        rot_matrix(rot_tel, orientation_telescope[i])
+
+        tod_det[i] += compute_signal_generic_hwp_for_one_sample(
+            Stokes=vec_S,
+            Vpol=polarimeter,
+            Rhwp=rot_hwp,
+            Mhwp=mueller_hwp,
+            Rtel=rot_tel,
         )
 
 
 def scan_map(
     tod,
     pointings,
-    hwp_angle,
-    maps: Union[np.ndarray, Dict[str, np.ndarray]],
-    input_names,
+    maps: Dict[str, np.ndarray],
+    pol_angle_detectors: Union[np.ndarray, None] = None,
+    pol_eff_detectors: Union[np.ndarray, None] = None,
+    hwp_angle: Union[np.ndarray, None] = None,
+    mueller_hwp: Union[np.ndarray, None] = None,
+    input_names: Union[str, None] = None,
     input_map_in_galactic: bool = True,
     interpolation: Union[str, None] = "",
 ):
@@ -55,17 +127,22 @@ def scan_map(
     interpolation, "linear" for linear interpolation)
     """
 
+    n_detectors = tod.shape[0]
+
     if type(pointings) is np.ndarray:
         assert tod.shape == pointings.shape[0:2]
 
-    for detector_idx in range(tod.shape[0]):
+    if pol_angle_detectors is None:
+        pol_angle_detectors = np.zeros(n_detectors)
+
+    if pol_eff_detectors is None:
+        pol_eff_detectors = np.ones(n_detectors)
+
+    for detector_idx in range(n_detectors):
         if type(pointings) is np.ndarray:
             curr_pointings_det = pointings[detector_idx, :, :]
         else:
             curr_pointings_det, hwp_angle = pointings(detector_idx)
-
-        if hwp_angle is None:
-            hwp_angle = 0
 
         if input_map_in_galactic:
             curr_pointings_det = rotate_coordinates_e2g(curr_pointings_det)
@@ -80,35 +157,64 @@ def scan_map(
         if interpolation in ["", None]:
             hpx = Healpix_Base(nside, "RING")
             pixel_ind_det = hpx.ang2pix(curr_pointings_det[:, 0:2])
-
-            scan_map_for_one_detector(
-                tod_det=tod[detector_idx],
-                input_T=maps_det[0, pixel_ind_det],
-                input_Q=maps_det[1, pixel_ind_det],
-                input_U=maps_det[2, pixel_ind_det],
-                pol_angle_det=curr_pointings_det[:, 2] + 2 * hwp_angle,
-            )
-
+            input_T = maps_det[0, pixel_ind_det]
+            input_Q = maps_det[1, pixel_ind_det]
+            input_U = maps_det[2, pixel_ind_det]
         elif interpolation == "linear":
-            scan_map_for_one_detector(
-                tod_det=tod[detector_idx],
-                input_T=hp.get_interp_val(
-                    maps_det[0, :], curr_pointings_det[:, 0], curr_pointings_det[:, 1]
-                ),
-                input_Q=hp.get_interp_val(
-                    maps_det[1, :], curr_pointings_det[:, 0], curr_pointings_det[:, 1]
-                ),
-                input_U=hp.get_interp_val(
-                    maps_det[2, :], curr_pointings_det[:, 0], curr_pointings_det[:, 1]
-                ),
-                pol_angle_det=curr_pointings_det[:, 2] + 2 * hwp_angle,
+            input_T = hp.get_interp_val(
+                maps_det[0, :], curr_pointings_det[:, 0], curr_pointings_det[:, 1]
             )
-
+            input_Q = hp.get_interp_val(
+                maps_det[1, :], curr_pointings_det[:, 0], curr_pointings_det[:, 1]
+            )
+            input_U = hp.get_interp_val(
+                maps_det[2, :], curr_pointings_det[:, 0], curr_pointings_det[:, 1]
+            )
         else:
             raise ValueError(
                 "Wrong value for interpolation. It should be one of the following:\n"
                 + '- "" for no interpolation\n'
                 + '- "linear" for linear interpolation\n'
+            )
+
+        if (mueller_hwp[detector_idx] is None) or (
+            (mueller_hwp[detector_idx] == mueller_ideal_hwp).all()
+        ):
+            # With HWP implements:
+            # (T + Q ρ Cos[2 (2 α - θ + ψ])] + U ρ Sin[2 (2 α - θ + ψ)])
+            # without
+            # (T + Q ρ Cos[2 (θ + ψ])] + U ρ Sin[2 (θ + ψ)])
+            # ρ: polarization efficiency
+            # θ: polarization angle
+            # ψ: angle of the telescope
+            # α: HWP angle
+            scan_map_for_one_detector(
+                tod_det=tod[detector_idx],
+                input_T=input_T,
+                input_Q=input_Q,
+                input_U=input_U,
+                pol_angle_det=(
+                    pol_angle_detectors[detector_idx] + curr_pointings_det[:, 2]
+                    if mueller_hwp[detector_idx] is None
+                    else 2 * hwp_angle
+                    - pol_angle_detectors[detector_idx]
+                    + curr_pointings_det[:, 2]
+                ),
+                pol_eff_det=pol_eff_detectors[detector_idx],
+            )
+        else:
+            # This implements:
+            # (1,0,0,0) x Mpol(ρ) x Rpol(θ) x Rhwp(α)^T x Mhwp x Rhwp(α) x Rtel(ψ) x Stokes
+            scan_map_generic_hwp_for_one_detector(
+                tod_det=tod[detector_idx],
+                input_T=input_T,
+                input_Q=input_Q,
+                input_U=input_U,
+                orientation_telescope=curr_pointings_det[:, 2],
+                pol_angle_det=pol_angle_detectors[detector_idx],
+                pol_eff_det=pol_eff_detectors[detector_idx],
+                hwp_angle=hwp_angle,
+                mueller_hwp=mueller_hwp[detector_idx],
             )
 
 
@@ -218,9 +324,16 @@ def scan_map_in_observations(
             input_names = None
 
         if hwp is None:
-            if hasattr(cur_obs, "hwp_angle"):
-                hwp_angle = cur_obs.hwp_angle
+            if cur_obs.has_hwp:
+                if hasattr(cur_obs, "hwp_angle"):
+                    hwp_angle = cur_obs.hwp_angle
+                else:
+                    hwp_angle = cur_obs.get_pointings()[1]
             else:
+                assert all(m is None for m in cur_obs.mueller_hwp), (
+                    "Detectors have been initialized with a mueller_hwp,"
+                    "but no HWP is either passed or initilized in the pointing"
+                )
                 hwp_angle = None
         else:
             if type(cur_ptg) is np.ndarray:
@@ -233,8 +346,11 @@ def scan_map_in_observations(
         scan_map(
             tod=getattr(cur_obs, component),
             pointings=cur_ptg,
-            hwp_angle=hwp_angle,
             maps=maps,
+            pol_angle_detectors=cur_obs.pol_angle_rad,
+            pol_eff_detectors=cur_obs.pol_efficiency,
+            hwp_angle=hwp_angle,
+            mueller_hwp=cur_obs.mueller_hwp,
             input_names=input_names,
             input_map_in_galactic=input_map_in_galactic,
             interpolation=interpolation,